fix tests compilation

+++ util/bigint/src/hash.rs
@@ -512,7 +512,7 @@ pub type H256FastSet = HashSet<H256, BuildHasherDefault<PlainHasher>>;
 #[cfg(test)]
 mod tests {
 	use hash::*;
-	use uint::*;
+	use bigint::*;
 	use std::str::FromStr;
 
 	#[test]
